revert: merge api and web server into one but keep different Tornado applications
